backend/requirements-rpi.txt
APScheduler==3.9.1
asgiref==3.5.2
colorzero==2.0
Django==4.1.2
django-cors-headers==3.13.0
django-ninja==0.17.0
et-xmlfile==1.1.0
gpiozero==1.6.2
gunicorn==20.1.0
numpy==1.23.4
openpyxl==3.1.1
pandas==1.5.1
pigpio==1.78
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
RPi.GPIO==0.7.1
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
